commit with relation ship

+++ DataTransferObject/Domain/Master/MArmedType.cs
@@ -12,7 +12,7 @@ namespace DataTransferObject.Domain.Master
     public class MArmedType:Common
     {
         [Key]
-        public int ArmedId { get; set; }
+        public short ArmedId { get; set; }
         
         [Required(ErrorMessage = "required!")]
         [RegularExpression("^[a-zA-Z0-9 ]*$", ErrorMessage = "Only Alphabets and Numbers allowed.")]
